+++ tests/introspection/ApiClassPtrTest.cs
@@ -17,9 +17,8 @@ using System.Runtime.CompilerServices;
 using Foundation;
 using ObjCRuntime;
 
-#if !NET
-using NativeHandle = System.IntPtr;
-#endif
+// Disable until we get around to enable + fix any issues.
+#nullable disable
 
 namespace Introspection {
