+++ tests/introspection/ApiSignatureTest.cs
@@ -31,18 +31,13 @@ using System.Linq;
 using Foundation;
 using ObjCRuntime;
 
-#if !NET
-using NativeHandle = System.IntPtr;
-#endif
+// Disable until we get around to enable + fix any issues.
+#nullable disable
 
 namespace Introspection {
 
 	public abstract class ApiSignatureTest : ApiBaseTest {
-#if NET
 		const string NFloatTypeName = "System.Runtime.InteropServices.NFloat";
-#else
-		const string NFloatTypeName = "System.nfloat";
-#endif
 		[DllImport ("/usr/lib/libobjc.dylib")]
 		// note: the returned string is not ours to free
 		static extern IntPtr objc_getClass (string name);
@@ -869,10 +864,8 @@ namespace Introspection {
 				}
 			}
 
-#if NET
 			if (TestRuntime.HasOSPlatformAttributeForCurrentPlatform<UnsupportedOSPlatformAttribute> (mi))
 				return true;
-#endif
 
 			return false;
 		}
