+++ tests/introspection/ApiWeakPropertyTest.cs
@@ -8,6 +8,9 @@ using NUnit.Framework;
 using Foundation;
 using ObjCRuntime;
 
+// Disable until we get around to enable + fix any issues.
+#nullable disable
+
 namespace Introspection {
 	[Preserve (AllMembers = true)]
 	public abstract class ApiWeakPropertyTest : ApiBaseTest {
