+++ tests/introspection/Assets.xcassets/AppIcons.appiconset/icon-app-57.png
