+++ tests/introspection/Assets.xcassets/AppIcons.appiconset/icon-app-60@2x.png
