+++ tests/introspection/Assets.xcassets/AppIcons.appiconset/icon-app-72@2x.png
