+++ tests/introspection/Assets.xcassets/AppIcons.appiconset/icon-app-76@2x.png
