+++ tests/introspection/Assets.xcassets/AppIcons.appiconset/icon-app-83.5@2x.png
