+++ tests/introspection/CoreSelectorTest.cs
@@ -14,6 +14,9 @@ using NUnit.Framework;
 using Foundation;
 using ObjCRuntime;
 
+// Disable until we get around to enable + fix any issues.
+#nullable disable
+
 namespace Introspection {
 
 	public abstract class CoreSelectorTest : ApiSelectorTest {
