+++ tests/introspection/Info.plist
