+++ tests/introspection/LaunchScreen.storyboard
