--- tests/introspection/Mac/introspection-mac.csproj
@@ -1,159 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">x86</Platform>
-    <ProjectGuid>{E0E84B58-C333-1BF3-D1EC-42EE3EA0D071}</ProjectGuid>
-    <ProjectTypeGuids>{A3F8F2AB-B479-4A4A-A458-A89E7DC349F1};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
-    <OutputType>Exe</OutputType>
-    <RootNamespace>introspection</RootNamespace>
-    <AssemblyName>introspection</AssemblyName>
-    <MonoMacResourcePrefix>Resources</MonoMacResourcePrefix>
-    <TargetFrameworkVersion>v2.0</TargetFrameworkVersion>
-    <IntermediateOutputPath>obj\$(Platform)\$(Configuration)</IntermediateOutputPath>
-    <TargetFrameworkIdentifier>Xamarin.Mac</TargetFrameworkIdentifier>
-    <XamMacArch>x86_64</XamMacArch>
-    <DefineConstants></DefineConstants>
-    <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
-    <LangVersion>latest</LangVersion>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|x86' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\x86\Debug</OutputPath>
-    <DefineConstants>DEBUG;MONOMAC;$(DefineConstants)</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>0</WarningLevel>
-    <EnableCodeSigning>true</EnableCodeSigning>
-    <CodeSigningKey>Mac Developer</CodeSigningKey>
-    <CreatePackage>false</CreatePackage>
-    <EnablePackageSigning>false</EnablePackageSigning>
-    <IncludeMonoRuntime>false</IncludeMonoRuntime>
-    <UseSGen>false</UseSGen>
-    <PlatformTarget>x86</PlatformTarget>
-    <PackageSigningKey>3rd Party Mac Developer Installer</PackageSigningKey>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|x86' ">
-    <Optimize>true</Optimize>
-    <OutputPath>bin\x86\Release</OutputPath>
-    <DefineConstants>MONOMAC;$(DefineConstants)</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>0</WarningLevel>
-    <EnableCodeSigning>true</EnableCodeSigning>
-    <CodeSigningKey>Developer ID Application</CodeSigningKey>
-    <CreatePackage>true</CreatePackage>
-    <EnablePackageSigning>false</EnablePackageSigning>
-    <IncludeMonoRuntime>true</IncludeMonoRuntime>
-    <UseSGen>false</UseSGen>
-    <LinkMode>SdkOnly</LinkMode>
-    <PlatformTarget>x86</PlatformTarget>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-    <Reference Include="System.Data" />
-    <Reference Include="Xamarin.Mac" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Resources\" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="MacApiCtorInitTest.cs" />
-    <Compile Include="MacApiFieldTest.cs" />
-    <Compile Include="MacApiSelectorTest.cs" />
-    <Compile Include="MacApiProtocolTest.cs" />
-    <Compile Include="MacApiSignatureTest.cs" />
-    <Compile Include="MacCoreImageFiltersTest.cs" />
-    <Compile Include="MacApiPInvokeTest.cs" />
-    <Compile Include="MacApiWeakPropertyTest.cs" />
-    <Compile Include="..\ApiBaseTest.cs">
-      <Link>ApiBaseTest.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiClassPtrTest.cs">
-      <Link>ApiClassPtrTest.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiCoreImageFiltersTest.cs">
-      <Link>ApiCoreImageFiltersTest.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiCtorInitTest.cs">
-      <Link>ApiCtorInitTest.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiFieldTest.cs">
-      <Link>ApiFieldTest.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiPInvokeTest.cs">
-      <Link>ApiPInvokeTest.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiProtocolTest.cs">
-      <Link>ApiProtocolTest.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiSelectorTest.cs">
-      <Link>ApiSelectorTest.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiSignatureTest.cs">
-      <Link>ApiSignatureTest.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiStructTest.cs">
-      <Link>ApiStructTest.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiTypoTest.cs">
-      <Link>ApiTypoTest.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiWeakPropertyTest.cs">
-      <Link>ApiWeakPropertyTest.cs</Link>
-    </Compile>
-    <Compile Include="..\CoreSelectorTest.cs">
-      <Link>CoreSelectorTest.cs</Link>
-    </Compile>
-    <Compile Include="..\EnvironmentVariable.cs">
-      <Link>EnvironmentVariable.cs</Link>
-    </Compile>
-    <Compile Include="..\..\common\PlatformInfo.cs">
-      <Link>PlatformInfo.cs</Link>
-    </Compile>
-    <Compile Include="..\..\common\mac\MacMain.cs">
-      <Link>MacMain.cs</Link>
-    </Compile>
-    <Compile Include="..\..\common\mac\Mac.cs">
-      <Link>Mac.cs</Link>
-    </Compile>
-    <Compile Include="..\..\common\TestRuntime.cs">
-      <Link>TestRuntime.cs</Link>
-    </Compile>
-    <Compile Include="MacApiTypoTest.cs" />
-    <Compile Include="..\ApiAvailabilityTest.cs">
-      <Link>ApiAvailabilityTest.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\tools\common\ApplePlatform.cs">
-      <Link>ApplePlatform.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\tools\common\Frameworks.cs">
-      <Link>Frameworks.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\tools\common\SdkVersions.cs">
-      <Link>SdkVersions.cs</Link>
-    </Compile>
-    <Compile Include="..\..\common\TestAssemblyLoader.cs">
-      <Link>TestAssemblyLoader.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiFrameworkTest.cs">
-      <Link>ApiFrameworkTest.cs</Link>
-    </Compile>
-    <Compile Include="..\ApiTypeTest.cs">
-      <Link>ApiTypeTest.cs</Link>
-    </Compile>
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\external\Touch.Unit\Touch.Client\macOS\mobile\Touch.Client-macOS-mobile.csproj">
-      <Project>{88A8A1AC-0829-4C98-8F4A-9FC23DC42A06}</Project>
-      <Name>Touch.Client-macOS-mobile</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="Info.plist" />
-  </ItemGroup>
-  <Import Project="$(MSBuildExtensionsPath)\Xamarin\Mac\Xamarin.Mac.CSharp.targets" />
-</Project>
-
