+++ tests/introspection/MacApiCtorInitTest.cs
@@ -36,11 +36,6 @@ namespace Introspection {
 		protected override bool Skip (Type type)
 		{
 			switch (type.FullName) {
-#if !NET
-			case "AppKit.NSDraggingInfo":
-			case "MonoMac.AppKit.NSDraggingInfo": // binding mistakes.
-				return true;
-#endif
 			// Random failures on build machine
 			case "QuickLookUI.QLPreviewPanel":
 			case "MonoMac.QuickLookUI.QLPreviewPanel":
