+++ tests/introspection/MacApiTypoTest.cs
@@ -9,7 +9,7 @@ namespace Introspection {
 
 	[TestFixture]
 	public class MacApiTypoTest : ApiTypoTest {
-		NSSpellChecker checker;
+		NSSpellChecker? checker;
 
 		public override void TypoTest ()
 		{
@@ -23,20 +23,10 @@ namespace Introspection {
 		{
 			var checkRange = new NSRange (0, txt.Length);
 			nint wordCount;
-			var typoRange = checker.CheckSpelling (txt, 0, "en_US", false, 0, out wordCount);
+			var typoRange = checker!.CheckSpelling (txt, 0, "en_US", false, 0, out wordCount);
 			if (typoRange.Length == 0)
 				return String.Empty;
 			return txt.Substring ((int) typoRange.Location, (int) typoRange.Length);
 		}
-
-#if !NET
-		public override bool Skip (Type baseType, string typo)
-		{
-			if (baseType == typeof (NSSpellCheckerCanidates))
-				return true;
-
-			return base.Skip (baseType, typo);
-		}
-#endif
 	}
 }
