+++ tests/introspection/MacApiWeakPropertyTest.cs
