+++ tests/introspection/MacCoreImageFiltersTest.cs
