+++ tests/introspection/README.md
@@ -1,11 +1,11 @@
 # Introspection Tests
 
-Introspection tests are executed on target (both simulator and device for
-iOS) or a specific version of OSX. The application proceed to analyze itself
-using:
+Introspection tests are executed on specific target (both simulator and device
+for iOS and tvOS) or a specific version of macOS or Mac Catalyst. The
+application proceed to analyze itself using:
 
 * `System.Reflection` for managed code; and
-* the ObjectiveC runtime library for native code
+* the Objective-C runtime library for native code
 
 and compare the results. E.g. if using .NET reflection it can see a binding
 for a `NSBundle` type then it should be able to find a native `NSBundle` 
@@ -13,7 +13,7 @@ type using the ObjC runtime functions. Otherwise an error is raised...
 
 Since the application analyze itself it must contains everything we wish
 to test. That's why the introspection tests needs to be built with the
-managed linker disable, i.e. **"Don't link"**.
+managed linker disabled, i.e. **"Don't link"**.
 
 Pros
