+++ tests/introspection/iOSApiFieldTest.cs
@@ -14,6 +14,9 @@ using ObjCRuntime;
 using UIKit;
 using NUnit.Framework;
 
+// Disable until we get around to enable + fix any issues.
+#nullable disable
+
 namespace Introspection {
 
 	[TestFixture]
@@ -166,10 +169,6 @@ namespace Introspection {
 			// Xcode 12.2 Beta 1 does not ship this but it is available in Xcode 12.0...
 			case "HKMetadataKeyBarometricPressure":
 				return true;
-#if __WATCHOS__
-			case "AVCaptureLensPositionCurrent": // looks like this was bound by mistake in watchOS
-				return true;
-#endif
 			default:
 				return false;
 			}
