+++ tests/introspection/iOSApiTypoTest.cs
@@ -9,28 +9,19 @@ namespace Introspection {
 
 	[TestFixture]
 	public class iOSApiTypoTest : ApiTypoTest {
-#if !__WATCHOS__
 		UITextChecker checker = new UITextChecker ();
-#endif
 
 		public override string GetTypo (string txt)
 		{
-#if __WATCHOS__
-			return string.Empty;
-#else
 			var checkRange = new NSRange (0, txt.Length);
 			var typoRange = checker.RangeOfMisspelledWordInString (txt, checkRange, checkRange.Location, false, "en_US");
 			if (typoRange.Length == 0)
 				return String.Empty;
 			return txt.Substring ((int) typoRange.Location, (int) typoRange.Length);
-#endif
 		}
 
 		public override void TypoTest ()
 		{
-#if __WATCHOS__
-			Assert.Ignore ("Need to find alternative for UITextChecker on WatchOS.");
-#else
 			// the dictionary used by iOS varies with versions and
 			// we don't want to maintain special cases for each version
 			var sdk = new Version (Constants.SdkVersion);
@@ -42,7 +33,6 @@ namespace Introspection {
 			TestRuntime.AssertSimulatorOrDesktop ("Typos only detected on simulator");
 
 			base.TypoTest ();
-#endif
 		}
 	}
 }
