+++ tests/introspection/iOSApiWeakPropertyTest.cs
@@ -5,6 +5,9 @@ using Foundation;
 
 using NUnit.Framework;
 
+// Disable until we get around to enable + fix any issues.
+#nullable disable
+
 namespace Introspection {
 
 	[TestFixture]
